VDI.detach: on failure to 'losetup -d', retry forever
These failures appear inevitable because other services on the system
(possibly the kernel?) take a passing interest in the device and cause
the 'losetup -d' to fail occasionally.

The VDI.detach API must not fail, so we retry forever. It should succeed
eventually because this is our loop device, no-one else should take
a permanent interest in it.

Fixes #4

Signed-off-by: David Scott <dave.scott@eu.citrix.com>