Handle both 'raw' and 'vhd' format files.
The config file specifies a default format, which will be applied
for new disks when the SR is attached unless an override is provided
in the device_config.

When a disk is created we first check to see whether an override
has been specified (sm-config:type=(vhd|raw)) and if not, we fall
back to the default set at SR attach time.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>