Make flexible tapdisk datapath
Renamed the former 'vhd' datapath to 'tapdisk' and created symbolic
links for 'raw+file' and 'vhd+file' pointing to tapdisk. This new
datapath code looks at the url scheme to determine whether an image is
raw or vhd and invokes tapdisk accordingly.

Also fixed the Makefile to create default symbolic links. The former
datapath implementation for 'raw+file' was renamed to 'loop+blkback'.

Signed-off-by: Felipe Franciosi <felipe@paradoxo.org>